more variable size adjustments.

+++ contracts/SocialToken/LongTailSocialToken.sol
@@ -43,9 +43,6 @@ contract LongTailSocialToken is ISocialToken, ERC777 {
     uint private linearInterestBonus;
     uint private quadraticInterestBonus;
 
-    // uint64 private nothing;
-    // uint64 private alsonothing;
-
     // TODO: mint tokens
     constructor(address manager_, address[] memory defaultOperators_) 
         ERC777("Long Tail Social Token", "LTST", defaultOperators_) {
@@ -248,7 +245,7 @@ contract LongTailSocialToken is ISocialToken, ERC777 {
         return votingPower;
     }
 
-    function calculateInterest(uint64 start, uint64 end, uint64 interestRate, uint256 principal) public view returns(bool, uint256) {
+    function calculateInterest(uint256 start, uint256 end, uint256 interestRate, uint256 principal) public view returns(bool, uint256) {
         uint256 halfStakeLength = (end - start) / 2;
         uint256 timeStaked = getCurrentDay() - start;
         uint256 payoff = _fullInterest(end - start, interestRate, principal);
@@ -260,7 +257,7 @@ contract LongTailSocialToken is ISocialToken, ERC777 {
         }
     }
 
-    function _votingWeight(uint64 start, uint64 end, uint256 currentDay) private pure returns(uint256){
+    function _votingWeight(uint256 start, uint256 end, uint256 currentDay) private pure returns(uint256){
         if (currentDay - start <= (end - start) / 2) {
             return (currentDay - start) * 2;
         }
@@ -269,12 +266,12 @@ contract LongTailSocialToken is ISocialToken, ERC777 {
         }
     }
 
-    function _fullInterest(uint64 duration, uint64 interestRate, uint256 principal) private pure returns(uint256) {
+    function _fullInterest(uint256 duration, uint256 interestRate, uint256 principal) private pure returns(uint256) {
         return (interestRate * duration * principal) / type(uint64).max;
     }
 
-    function calculateInterestRate(address account, uint64 numberOfDays) public view returns(uint64) {
-        uint256 interest = baseInterestRate + uint256(linearInterestBonus * numberOfDays) + uint256(quadraticInterestBonus * numberOfDays * numberOfDays) + uint256(manager.getNftContract().interestBonus(account));
+    function calculateInterestRate(address account, uint256 numberOfDays) public view returns(uint64) {
+        uint256 interest = baseInterestRate + linearInterestBonus * numberOfDays + quadraticInterestBonus * numberOfDays * numberOfDays + manager.getNftContract().interestBonus(account);
         // cap the value at what can be held in a uint64 and downcast it into a uint32
         return interest > type(uint64).max ? type(uint64).max : uint64(interest);
     }
